File manager scalability patch
