+++ managers/file_manager/parsers/__init__.py
@@ -1,4 +1,4 @@
-from .parser_base import IParser, IParserFactory, read_stream, write_stream
+from .parser_base import IParser, IParserFactory
 from .parser_factory import ParserFactory
 from .parser import Parser
